Create Lines

Create Lines that are compatible with CCAMLR online GIS

create_Lines(InputFile, OutputName, Buffer = 0, Separate = 0, Densify = 1,
  Clip = 0)

Arguments

InputFile
currently, the name of the input data file in quotes e.g. "DataFile.csv", but I want to change this so you can choose between R dataframe or csv file

OutputName
currently, the name of the output shapefile in quotes e.g."MyShape" but I want to change this to "ROBJECT" is temporarily stored in work environment, "RDATA" will write an .rda file or "SHAPEFILE" with write an ESRI Shapefile

Buffer
is the value in nautical miles to apply to the line. The default value is 0, assuming no Buffer

Separate
currently, If the data file contains several items (e.g. several fishing lines) you may wish to produce one shapefile per item to display them separately on a map. In such a case, set Separate to 1. I want to remove this as I think you can manage this at the input end.

Densify
is set to 1 as a default, which will add additional points between points of equal latitude when data are projected. If set to 0 then no additional points will be added

Clip
"Coast_Low" will clip a line of polygon that intersect with the coastline to remove the land and keep only the ocean area, "Coast_Medium" is a higher resolution coastline is also provided with the package, the default is set to 0 which assumes no clipping is required